# docs/swagger.yaml
basePath: /api/v1
definitions:
  http.AskRequest:
    properties:
      question:
        type: string
    required:
    - question
    type: object
...
  response.ErrorData:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  response.StandardResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorData'
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
  usecase.CreateAnnouncementInput:
    properties:
      app_id:
        type: integer
      ends_at:
        type: string
      is_active:
        type: boolean
      judul:
        type: string
      konten:
        type: string
      starts_at:
        type: string
      tipe:
        description: info, warning, maintenance, urgent
        type: string
    required:
    - judul
    - konten
    type: object
...
  usecase.UpdateAnnouncementInput:
    properties:
      app_id:
        type: integer
      ends_at:
        type: string
      is_active:
        type: boolean
      judul:
        type: string
      konten:
        type: string
      starts_at:
        type: string
      tipe:
        type: string
    required:
    - judul
    - konten
    type: object
...
host: localhost:8080
info:
  contact:
    email: bps1200@bps.go.id
    name: Tim Inovasi BPS Provinsi Sumatera Utara
    url: https://sumut.bps.go.id
  description: Backend RESTful API untuk Portal Aplikasi Terintegrasi BPS Provinsi
    Sumatera Utara.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sumut.bps.go.id
  title: Portal Aplikasi BPS Provinsi Sumatera Utara (ALUSI) API
  version: 1.2.0
paths:
  /admin/analytics/disruptions:
    get:
      description: Mengambil riwayat insiden kendala operasional dan persentase uptime
        SLA per layanan (Admin)
      parameters:
      - description: Rentang hari ke belakang (default 30)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Rekapitulasi gangguan layanan dan SLA
      tags:
      - Admin - Analytics
  /admin/analytics/summary:
    get:
      description: Mengambil agregasi total pengguna, total aplikasi aktif, total
        klik, serta DAU dan MAU (Admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Ringkasan metrik eksekutif (KPI Rollup)
      tags:
      - Admin - Analytics
  /admin/analytics/top-apps:
    get:
      description: Mengambil daftar peringkat aplikasi terpopuler berdasarkan jumlah
        klik (Admin)
      parameters:
      - description: Rentang hari ke belakang (default 30)
        in: query
        name: days
        type: integer
      - description: Jumlah item ranking (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Ranking aplikasi paling sering diakses
      tags:
      - Admin - Analytics
  /admin/analytics/trends:
    get:
      description: Mengambil data tren grafik klik harian dan jumlah pengguna aktif
        unik (Admin)
      parameters:
      - description: Rentang hari ke belakang (default 30)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Tren trafik dan aktivitas harian
      tags:
      - Admin - Analytics
  /admin/announcements:
    get:
      description: Mengambil seluruh riwayat pengumuman baik yang aktif maupun nonaktif
        (Admin)
      parameters:
      - description: Halaman (default 1)
        in: query
        name: page
        type: integer
      - description: Jumlah item per halaman (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Daftar seluruh pengumuman (Admin)
      tags:
      - Admin - Announcements
    post:
      consumes:
      - application/json
      description: Membuat pengumuman broadcast atau banner pemeliharaan baru (Admin)
      parameters:
      - description: Data pengumuman baru
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateAnnouncementInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Buat pengumuman baru
      tags:
      - Admin - Announcements
  /admin/announcements/{id}:
    delete:
      description: Menghapus pengumuman dari sistem (Admin)
      parameters:
      - description: ID Pengumuman
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Hapus pengumuman
      tags:
      - Admin - Announcements
    put:
      consumes:
      - application/json
      description: Memperbarui judul, pesan, tipe, atau masa tayang pengumuman (Admin)
      parameters:
      - description: ID Pengumuman
        in: path
        name: id
        required: true
        type: integer
      - description: Data perubahan pengumuman
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecase.UpdateAnnouncementInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Perbarui pengumuman
      tags:
      - Admin - Announcements
...
  /ai/ask:
    post:
      consumes:
      - application/json
      description: Memproses pertanyaan natural bahasa Indonesia untuk merekomendasikan
        aplikasi BPS yang relevan
      parameters:
      - description: Pertanyaan pengguna
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/http.AskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Tanya Asisten AI & Rekomendasi Cerdas
      tags:
      - AI Assistant
  /announcements:
    get:
      description: Mengambil pengumuman broadcast sistem atau banner pemeliharaan
        spesifik aplikasi
      parameters:
      - description: Filter spesifik ID aplikasi (opsional)
        in: query
        name: app_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Daftar pengumuman / banner aktif
      tags:
      - Announcements
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan JWT token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"